trying to find a way to conditionnaly render pages based on user roles, also morganized the files better
