+++ petclinic-frontend/src/features/customers/components/UpdateCustomerForm.css
@@ -1,4 +1,4 @@
-.profile-edit {
+.update-customer-form {
     max-width: 600px;
     margin: 50px auto;
     padding: 20px;
@@ -8,25 +8,25 @@
     box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
 }
 
-.profile-edit h1 {
+.update-customer-form h1 {
     text-align: center;
     margin-bottom: 20px;
     color: #333;
     font-family: 'Arial', sans-serif;
 }
 
-.profile-edit form {
+.update-customer-form form {
     display: flex;
     flex-direction: column;
 }
 
-.profile-edit label {
+.update-customer-form label {
     margin-bottom: 15px;
     font-family: 'Arial', sans-serif;
     color: #555;
 }
 
-.profile-edit input {
+.update-customer-form input {
     padding: 10px;
     margin-top: 5px;
     border: 1px solid #ccc;
@@ -35,7 +35,7 @@
     font-size: 16px;
 }
 
-.profile-edit button {
+.update-customer-form button {
     padding: 12px;
     background-color: #28a745;
     color: white;
@@ -48,7 +48,7 @@
     transition: background-color 0.3s ease;
 }
 
-.profile-edit button:hover {
+.update-customer-form button:hover {
     background-color: #218838;
 }
 
@@ -59,7 +59,7 @@
     font-family: 'Arial', sans-serif;
 }
 
-.profile-edit input:focus {
+.update-customer-form input:focus {
     border-color: #28a745;
     box-shadow: 0 0 5px rgba(40, 167, 69, 0.5);
     outline: none;
