+++ petclinic-frontend/src/shared/models/Role.ts
@@ -1,0 +1,4 @@
+export interface Role {
+  id: number;
+  name: string;
+}
