+++ petclinic-frontend/src/shared/models/UserResponseModel.ts
@@ -1,6 +1,8 @@
+import { Role } from '@/shared/models/Role.ts';
+
 export interface UserResponseModel {
   username: string;
   email: string;
   userId: string;
-  roles: string;
+  roles: Set<Role>;
 }
